backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
PyJWT==2.9.0
cryptography==41.0.2
redis==5.0.5
requests==2.31.0
